[test] Store Service 가게 등록 단위 테스트 (#6)

+++ src/main/java/com/cakestation/backend/service/dto/request/CreateStoreDto.java
@@ -31,7 +31,7 @@ public class CreateStoreDto {
 
     private String kakaoMapUrl;
 
-    private Float score;
+    private Double score;
 
     public static Store toEntity(User user, CreateStoreDto createStoreDto){
         return Store.builder()
